oracle-poja-base/src/test/java/com/company/base/endpoint/rest/controller/HealthControllerTest.java
package com.company.base.endpoint.rest.controller;

import com.company.base.FacadeTest;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;

import static org.junit.jupiter.api.Assertions.assertEquals;

class HealthControllerTest extends FacadeTest {

  @Autowired
  HealthController healthController;

  @Test
  void ping() {
    assertEquals("pong", healthController.ping());
  }

  @Test
  void dummyTable() {
    var dummyTableEntries = healthController.dummyTable();
    assertEquals(1, dummyTableEntries.size());
    assertEquals("dummy-table-id-1", dummyTableEntries.get(0).getId());
  }
}